Refactor primary version of Sweep Private Key modal
