+++ src/__tests__/scenesReducer.test.js
@@ -61,9 +61,6 @@ test('initialState', () => {
       privateKeyModal: {
         error: null,
         isSweeping: false,
-        primaryModal: {
-          isActive: false
-        },
         secondaryModal: {
           isActive: false
         }
