--- src/actions/PrimaryModalActions.js
@@ -1,9 +1,0 @@
-// @flow
-
-export const activated = () => ({
-  type: 'PRIVATE_KEY_MODAL/PRIMARY_MODAL/ACTIVATED'
-})
-
-export const deactivated = () => ({
-  type: 'PRIVATE_KEY_MODAL/PRIMARY_MODAL/DEACTIVATED'
-})
