+++ src/actions/SecondaryModalActions.js
@@ -1,9 +1,9 @@
 // @flow
 
-export const activated = () => ({
+export const secondaryModalActivated = () => ({
   type: 'PRIVATE_KEY_MODAL/SECONDARY_MODAL/ACTIVATED'
 })
 
-export const deactivated = () => ({
+export const secondaryModalDeactivated = () => ({
   type: 'PRIVATE_KEY_MODAL/SECONDARY_MODAL/DEACTIVATED'
 })
