+++ src/components/modals/PrivateKeyModal.js
@@ -3,7 +3,6 @@
 import React, { Component } from 'react'
 import { View } from 'react-native'
 
-import { PrimaryModalConnector as PrimaryModal } from '../../connectors/PrimaryModalConnector.js'
 import { SecondaryModalConnector as SecondaryModal } from '../../connectors/SecondaryModalConnector.js'
 
 type Props = {
@@ -13,10 +12,9 @@ type Props = {
 }
 export class PrivateKeyModal extends Component<Props> {
   render () {
-    const { onPrivateKeyAccept, reset } = this.props
+    const { reset } = this.props
     return (
       <View>
-        <PrimaryModal onAccept={onPrivateKeyAccept} />
         <SecondaryModal onModalHide={reset} />
       </View>
     )
